requirements.txt
beautifulsoup4
discord.py
requests
aiohttp
asyncio
Brotli